Chat Bot/backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
pydantic-settings>=2.9.0
google-genai>=1.14.0
chromadb>=1.0.0
python-docx>=1.1.0
pandas>=2.2.0
openpyxl>=3.1.0
python-multipart>=0.0.20
tabulate>=0.9.0
